requirements.txt
hydra-core==1.3.2
joblib~=1.3.2
matplotlib~=3.9.1
more_itertools~=10.3.0
numpy==1.26.4
pandas~=2.2.3
# pydantic
git+https://github.com/samuelcolvin/pydantic.git@fbf8002cefd1ba7b263f63545fff2e48a378189d#egg=pydantic
scikit-learn~=1.5.2
scipy~=1.14.1
seaborn~=0.13.2
tabulate~=0.9.0
tables~=3.9.2
tqdm~=4.66.1